requirements.txt
requests==2.31.0
pandas==2.1.4
pyyaml==6.0.1
python-dotenv>=0.19.0  # 环境变量管理